Fixed #23 -- Simplified code but retained JSON parsing.
Removed unnecessary conditional check for element existence since it will
always be present in the template.
Kept JSON.parse() as it's necessary when using json_script because the
template tag serializes values to JSON format, wrapping even simple strings
in quotes. Without it, navigation would fail due to the browser attempting
to navigate to a URL with quotation marks.